Fix the nav items broken since gdpr nav edit
The main nav is set to a specific margin and width. Adding a new nav item causes this to shift.
Also the count of nav items changes. So I've updated the CSS to not count the nav items for the JOIN. 
Here is the block showing the JOIN button https://community.blackbaud.com/admin/builder/block/14583291188502